requirements.txt
streamlit
requests
beautifulsoup4
scikit-learn
python-dotenv
httpx
